+++ behaviours/messageparty.py
@@ -1,19 +1,14 @@
-import os, db
-import hikari, lightbulb
-
-plugin = lightbulb.Plugin("messageparty")
+import db
 
 """
 Kitti congratulates you for your efforts.
 """
 
 
-@plugin.listener(hikari.GuildMessageCreateEvent)
 async def main(event):
     if event.is_bot or not event.content:
         return
 
-    messageContent = event.content
     user_id = event.author_id
     cursor = db.cursor()
     cursor.execute(
@@ -41,7 +36,3 @@ async def main(event):
     elif message_count % 1000 == 0:
         response1 = f"""<a:partyblob:815938533470240799> <a:partyblob:815938533470240799> <a:partyblob:815938533470240799> {event.author.mention}, congratulations on sending {message_count_formatted} messages! <a:partyblob:815938533470240799> <a:partyblob:815938533470240799> <a:partyblob:815938533470240799>."""
         await event.message.respond(response1, user_mentions=True)
-
-
-def load(bot: lightbulb.BotApp):
-    bot.add_plugin(plugin)
