+++ commands/fortune.py
